+++ backend/src/main/java/com/yuxian/backend/entity/Product.java
@@ -6,6 +6,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Data;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 @Data 
 @Entity
@@ -13,14 +14,14 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-
     private Long id;
     private String category;
     private String name;
     private String origin;
     private LocalDate listDate;
     private String description;
-    private Double price;
+    private BigDecimal price; 
+    
     private Integer stock;
     private String imageUrl;
 }
